requirements.txt
requests
twine
pytest
coverage
pytest-cov
ansicolors
python-editor
